requirements.txt
torch
torchvision
timm
pytorch-lightning
numpy
opencv-python
scikit-learn
onnx
onnxruntime
datasets
imagehash
pillow
tqdm
tensorboard